changed BIGINT AUTO_INCREMENT to SERIAL for postgresql

+++ src/main/resources/db/migration/postgresql/R__insert_data.sql
@@ -1,4 +1,4 @@
-DELETE FROM APP_USER;
+DELETE FROM APP_USER WHERE USER_NAME = 'admin@example.org';
 
 INSERT INTO APP_USER (FIRST_NAME, LAST_NAME, USER_NAME, PASSWORD, ROLE) VALUES
        ('Admin', 'Administratorski', 'admin@example.org' , '$2a$10$1YFQa/2ctYg62/P4aXGKeOmB2snioaP4vDtbou8OMy64dFEbu1N9u', 'ROLE_ADMIN');
